backend/requirements.txt
fastapi
uvicorn
databricks-sql-connector
pydantic
python-dotenv
